Fix prevent issue - Unsigned compared against 0
- Since t/step can be negative number, min = static_cast<int>(t/step); and
if( min < 0) might cause unexpected behavior.

<Prevent message>
CID 436228 (#1 of 1): Unsigned compared against 0 (NO_EFFECT)
unsigned_compare: This less-than-zero comparison of an unsigned value is never true. min < 0U.
 95        if( min < 0)

+++ dali/internal/event/animation/linear-constrainer-impl.h
@@ -84,8 +84,8 @@ struct LinearConstraintFunctor
       float t = (( inputWrapped - mRange.x ) / ( mRange.y-mRange.x ));
 
       //Find min and max values and local t between them
-      size_t min(0);
-      size_t max(0);
+      int min(0);
+      int max(0);
       float tLocal(0.0f);
       if( mProgress.Size() < valueCount )
       {
